Move LROs and PagedResponse types into models as well.
