+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/key_client.hpp
@@ -10,9 +10,7 @@
 #pragma once
 
 #include "azure/keyvault/keys/key_client_models.hpp"
-#include "azure/keyvault/keys/key_client_operations.hpp"
 #include "azure/keyvault/keys/key_client_options.hpp"
-#include "azure/keyvault/keys/key_client_paged_responses.hpp"
 
 #include <functional>
 #include <vector>
