+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keyvault_keys.hpp
@@ -16,6 +16,4 @@
 #include "azure/keyvault/keys/dll_import_export.hpp"
 #include "azure/keyvault/keys/key_client.hpp"
 #include "azure/keyvault/keys/key_client_models.hpp"
-#include "azure/keyvault/keys/key_client_operations.hpp"
 #include "azure/keyvault/keys/key_client_options.hpp"
-#include "azure/keyvault/keys/key_client_paged_responses.hpp"
